+++ src/Modules/Payment/Explorer.Payment.API/Public/Author/ITourSaleService.cs
@@ -12,7 +12,7 @@ namespace Explorer.Payment.API.Public.Author
     public interface ITourSaleService
     {
         public Result<TourSaleDto> Create(TourSaleDto tourSaleDto, int userId);
-        public Result<TourSaleDto> Update(int id, TourSaleDto tourSaleDto);
+        public Result<TourSaleDto> Update(long id, TourSaleDto tourSaleDto);
         public Result<bool> Delete(long id);
         public Result<List<TourSaleDto>> GetAll();
         Result<PagedResult<TourSaleDto>> GetPaged(int page, int pageSize);
